+++ src/pytest_suite_timeout/plugin.py
@@ -1,7 +1,14 @@
 import time
 import pytest
 
-_expire_time = 0
+suite_timeout_key = pytest.StashKey[float]()
+
+
+SUITE_TIMEOUT_DESC = """
+Timeout in seconds for entire suite.  Default is None which
+means no timeout. Timeout is checked between tests, and will not interrupt a test 
+in progress. Can be specified as a float.
+"""
 
 def pytest_addoption(parser):
     parser.addoption(
@@ -10,19 +17,23 @@ def pytest_addoption(parser):
         dest="suite_timeout",
         default=None,
         type=float,
-        metavar="minutes",
-        help="Don't run for longer than this.")
+        metavar="SECONDS",
+        help=SUITE_TIMEOUT_DESC)
 
 
 def pytest_configure(config):
-    global _expire_time, _minutes
-    _minutes = config.getoption("--suite-timeout")
-    if _minutes:
-        _expire_time = time.time() + (_minutes * 60)
+    timeout = config.getoption("--suite-timeout")
+    if timeout is not None:
+        expire_time = time.time() + timeout
+    else:
+        expire_time = 0
+    config.stash[suite_timeout_key] = expire_time
 
 
-def pytest_runtest_logfinish(nodeid, location):
-    if _expire_time and _expire_time < time.time():
-        pytest.exit(f"suite-timeout: {_minutes} minutes exceeded",
-                    returncode=0)
-
+def pytest_runtest_makereport(item, call):
+    session = item.session
+    config = session.config
+    expire_time = config.stash[suite_timeout_key]
+    if expire_time and (expire_time < time.time()):
+        timeout = config.getoption("--suite-timeout")
+        session.shouldfail = f"suite-timeout: {timeout} sec exceeded"